requirements.txt
numpy
pandas
orjson
gpustat
torch
tensorboardX
transformers
datasets
accelerate
matplotlib
hydra-core
omegaconf
rouge_score
tqdm
matplotlib
einops
packaging
scipy
ninja
